Moved calculate_gradient_across_cell_corners.
Moved calculate_gradient_across_cell_corners to the raster_gradients
module.
